requirements.txt
django==3.2.12
djangorestframework
markdown
django-filter
mysqlclient
pillow
django-crispy-forms
django-import-export
django-reversion
django-formtools
django-guardian
future
httplib2
six
xlwt
xlsxwriter
coreapi